web: Move the players to internal/player
